--- a/cli/src/opts/forge.rs
+++ b/cli/src/opts/forge.rs
@@ -3,7 +3,9 @@ use clap::{Parser, Subcommand, ValueHint};
 use ethers::{solc::EvmVersion, types::Address};
 use std::{path::PathBuf, str::FromStr};
 
-use crate::cmd::{build::BuildArgs, create::CreateArgs, run::RunArgs, snapshot, test};
+use crate::cmd::{
+    build::BuildArgs, create::CreateArgs, remappings::RemappingArgs, run::RunArgs, snapshot, test,
+};
 
 #[derive(Debug, Parser)]
 pub struct Opts {