Co-authored with @brockelmore, converts unit & fuzz tests to be run in parallel.

This achieves a significant speedup, especially for fuzz tests.

Things in this PR:
1. Instead of passing EVM into the runner, pass `EvmOpts` (refactor EvmOpts to evm-adapters)
2. Instantiate an EVM on each test (via `new_sputnik_evm`, this tightly couples us to SputnikEVM a bit, but should be fine, we weren't using evmodin either way)
3. Use the `SharedBackend` (instead of `ForkMemoryBackend`), which is cheap to clone, but allows sharing state across forked tests, deduplicating work as much as possible
4. Instantiating an EVM per test lets us make our functions `&self` instead of `&mut self`, which makes parallelization with `rayon` very easy: we just replace `.iter()` with `par_iter()`.

Supersedes https://github.com/gakonst/foundry/pull/435
Also contains the fixes from https://github.com/gakonst/foundry/pull/430.


* [x] Benchmark
     * @brockelmore tested `spells-mainnet` from 2m37 go to 1m43 (~30% improv)
     * [`guni-lev`](https://github.com/hexonaut/guni-lev/) from to 33s to 27.7s (~20% improvement)
     * [`drai`](https://github.com/mds1/drai) from ~18s to ~9s (2x improv)
     * **[`solmate`](https://github.com/Rari-Capital/solmate) from 2m48 to 47s (3.5x)**
* [x] Figure out the todo-item in `evm.debug_calls()` in the CLI's `run.rs` (https://github.com/gakonst/foundry/pull/445)